requirements.txt
asgiref==3.5.1
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.3
distlib==0.3.4
Django==3.2.12
filelock==3.6.0
future==0.16.0
idna==3.3
jmespath==0.10.0
mysql-connector-python==8.0.29
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.20.1
pypiwin32==223
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1

